Renames proxy_decl to view_decl
