+++ core/include/scipp/core/view_decl.h
